+++ src/main/kotlin/io/github/jamalam/util/Constants.kt
@@ -1,0 +1,29 @@
+/*
+ * Copyright (C) 2022 Jamalam
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package io.github.jamalam.util
+
+/**
+ * @author  Jamalam360
+ */
+
+const val PINGUINO_PFP = "https://images-ext-2.discordapp.net/external/tM2ezTNgh6TK_9IW5eCGQLtuaarLJfjdRgJ3hmRQ5rs" +
+        "/%3Fsize%3D256/https/cdn.discordapp.com/avatars/896758540784500797/507601ac" +
+        "31f51ffc334fac125089f7ea.png"
+
+const val VERSION = "v0.7.0"
+const val DBL_URL = "https://top.gg/api/bots/896758540784500797/stats"